requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
addict==2.2.1             # via -r requirements.in, mmcv
albumentations==0.4.5     # via -r requirements.in
attrs==19.3.0             # via pytest
backcall==0.1.0           # via ipython
bbaug==0.4.2              # via -r requirements.in
cycler==0.10.0            # via matplotlib
cython==0.29.17           # via -r requirements.in
decorator==4.4.2          # via ipython, networkx, traitlets
ensemble-boxes==1.0.2     # via -r requirements.in
future==0.18.2            # via torch
imagecodecs==2020.2.18    # via tifffile
imageio==2.8.0            # via scikit-image
imgaug==0.2.6             # via albumentations, bbaug
importlib-metadata==1.6.0  # via pluggy, pytest
ipdb==0.13.2              # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.14.0           # via -r requirements.in, ipdb
iterative-stratification==0.1.6  # via -r requirements.in
jedi==0.17.0              # via ipython
joblib==0.16.0            # via scikit-learn
kiwisolver==1.2.0         # via matplotlib
loguru==0.5.0             # via -r requirements.in
matplotlib==3.2.1         # via -r requirements.in, scikit-image
mmcv==0.5.9               # via -r requirements.in
more-itertools==8.3.0     # via pytest
networkx==2.4             # via scikit-image
numpy==1.18.4             # via -r requirements.in, albumentations, ensemble-boxes, imagecodecs, imageio, imgaug, iterative-stratification, matplotlib, mmcv, opencv-python, pandas, pywavelets, scikit-image, scikit-learn, scipy, tifffile, torch, torchvision
opencv-python==4.2.0.34   # via albumentations, mmcv
packaging==20.3           # via pytest
pandas==1.0.3             # via -r requirements.in, ensemble-boxes
parso==0.7.0              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==6.2.2             # via -r requirements.in, imageio, scikit-image, torchvision
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.5     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.8.1                 # via pytest
pygments==2.6.1           # via ipython
pyparsing==2.4.7          # via matplotlib, packaging
pytest==5.4.2             # via -r requirements.in
python-dateutil==2.8.1    # via matplotlib, pandas
pytz==2020.1              # via pandas
pywavelets==1.1.1         # via scikit-image
pyyaml==5.3.1             # via albumentations, mmcv
scikit-image==0.17.2      # via imgaug
scikit-learn==0.23.2      # via iterative-stratification
scipy==1.4.1              # via albumentations, imgaug, iterative-stratification, scikit-image, scikit-learn
six==1.14.0               # via -r requirements.in, cycler, imgaug, packaging, python-dateutil, traitlets
terminaltables==3.1.0     # via -r requirements.in
threadpoolctl==2.1.0      # via scikit-learn
tifffile==2020.5.11       # via scikit-image
timm==0.1.30              # via -r requirements.in
torch==1.5.0              # via -r requirements.in, timm, torchvision
torchvision==0.6.0        # via -r requirements.in, timm
traitlets==4.3.3          # via ipython
wcwidth==0.1.9            # via prompt-toolkit, pytest
yapf==0.30.0              # via -r requirements.in, mmcv
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
